LayoutTests/imported/w3c/web-platform-tests/websockets/opening-handshake/001-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/invalid' failed: Invalid HTTP version string: FOO

PASS WebSockets: invalid handshake 
